+++ s202350103/src/main/java/com/choongang/s202350103/hrService/OrderService.java
@@ -8,7 +8,7 @@ public interface OrderService {
 	// boOrderList.jsp
 	int 		 countOrderrList();
 	List<Orderr> selectOrderrList(Orderr orderr);
-	Orderr 		 selectOrderr(int o_order_num);
+	Orderr 		 selectOrderr(long o_order_num);
